added special challenge group logic, challenge done button logic, fixing bugs

+++ src/shared/lib/firebase/storageServices.ts
@@ -1,5 +1,11 @@
 import storage from '@react-native-firebase/storage';
 
+interface FirebaseStorageError extends Error {
+  code: string;
+  message: string;
+  name: string;
+}
+
 export class StorageServices {
   storageRef = storage().ref();
 
@@ -8,15 +14,56 @@ export class StorageServices {
     this.storageRef = ref.child(folderName).child(fileName);
   }
 
+  async fileExists(): Promise<boolean> {
+    try {
+      // Attempt to get the download URL, which only succeeds if the file exists
+      await this.storageRef.getDownloadURL();
+      return true;
+    } catch (error) {
+      const storageError = error as FirebaseStorageError;
+
+      if (storageError.code === 'storage/object-not-found') {
+        return false;
+      }
+      throw error;
+    }
+  }
+
   async upload(file: string) {
-    await this.storageRef.putFile(file);
+    try {
+      const response = await this.storageRef.putFile(file);
+      console.log('File uploaded successfully:', response);
+      return response;
+    } catch (error) {
+      console.error('Upload error:', error);
+      throw error;
+    }
   }
 
   async download() {
-    return await this.storageRef.getDownloadURL();
+    try {
+      const url = await this.storageRef.getDownloadURL();
+      console.log('Download URL:', url);
+      return url;
+    } catch (error) {
+      console.error('Download error:', error);
+      throw error;
+    }
   }
 
-  async detete() {
-    await this.storageRef.delete();
+  async delete() {
+    try {
+      const exists = await this.fileExists();
+      if (!exists) {
+        console.log('No file exists at the specified reference.');
+        return;
+      }
+
+      await this.storageRef.delete();
+      console.log('File deleted successfully.');
+    } catch (error) {
+      console.error('Error deleting file:', error);
+      throw error;
+    }
   }
 }
